+++ wow_vanilla_messages/src/world/v1/v12/cmsg_activatetaxiexpress.rs
@@ -21,7 +21,7 @@ pub struct CMSG_ACTIVATETAXIEXPRESS {
 }
 
 impl ClientMessageWrite for CMSG_ACTIVATETAXIEXPRESS {
-    const OPCODE: u32 = 0x312;
+    const OPCODE: u16 = 0x312;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
