+++ wow_vanilla_messages/src/world/v1/v12/cmsg_area_spirit_healer_query.rs
@@ -19,7 +19,7 @@ pub struct CMSG_AREA_SPIRIT_HEALER_QUERY {
 }
 
 impl ClientMessageWrite for CMSG_AREA_SPIRIT_HEALER_QUERY {
-    const OPCODE: u32 = 0x2e2;
+    const OPCODE: u16 = 0x2e2;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
