+++ wow_vanilla_messages/src/world/v1/v12/cmsg_auction_place_bid.rs
@@ -21,7 +21,7 @@ pub struct CMSG_AUCTION_PLACE_BID {
 }
 
 impl ClientMessageWrite for CMSG_AUCTION_PLACE_BID {
-    const OPCODE: u32 = 0x25a;
+    const OPCODE: u16 = 0x25a;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
