+++ wow_vanilla_messages/src/world/v1/v12/cmsg_auction_remove_item.rs
@@ -20,7 +20,7 @@ pub struct CMSG_AUCTION_REMOVE_ITEM {
 }
 
 impl ClientMessageWrite for CMSG_AUCTION_REMOVE_ITEM {
-    const OPCODE: u32 = 0x257;
+    const OPCODE: u16 = 0x257;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
