+++ wow_vanilla_messages/src/world/v1/v12/cmsg_binder_activate.rs
@@ -19,7 +19,7 @@ pub struct CMSG_BINDER_ACTIVATE {
 }
 
 impl ClientMessageWrite for CMSG_BINDER_ACTIVATE {
-    const OPCODE: u32 = 0x1b5;
+    const OPCODE: u16 = 0x1b5;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
