+++ wow_vanilla_messages/src/world/v1/v12/cmsg_buy_item.rs
@@ -22,7 +22,7 @@ pub struct CMSG_BUY_ITEM {
 }
 
 impl ClientMessageWrite for CMSG_BUY_ITEM {
-    const OPCODE: u32 = 0x1a2;
+    const OPCODE: u16 = 0x1a2;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
