+++ wow_vanilla_messages/src/world/v1/v12/cmsg_cancel_aura.rs
@@ -18,7 +18,7 @@ pub struct CMSG_CANCEL_AURA {
 }
 
 impl ClientMessageWrite for CMSG_CANCEL_AURA {
-    const OPCODE: u32 = 0x136;
+    const OPCODE: u16 = 0x136;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
