+++ wow_vanilla_messages/src/world/v1/v12/cmsg_cancel_channelling.rs
@@ -18,7 +18,7 @@ pub struct CMSG_CANCEL_CHANNELLING {
 }
 
 impl ClientMessageWrite for CMSG_CANCEL_CHANNELLING {
-    const OPCODE: u32 = 0x13b;
+    const OPCODE: u16 = 0x13b;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
