+++ wow_vanilla_messages/src/world/v1/v12/cmsg_cancel_trade.rs
@@ -17,7 +17,7 @@ pub struct CMSG_CANCEL_TRADE {
 }
 
 impl ClientMessageWrite for CMSG_CANCEL_TRADE {
-    const OPCODE: u32 = 0x11c;
+    const OPCODE: u16 = 0x11c;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
