+++ wow_vanilla_messages/src/world/v1/v12/cmsg_channel_set_owner.rs
@@ -18,7 +18,7 @@ pub struct CMSG_CHANNEL_SET_OWNER {
 }
 
 impl ClientMessageWrite for CMSG_CHANNEL_SET_OWNER {
-    const OPCODE: u32 = 0x9d;
+    const OPCODE: u16 = 0x9d;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         self.size() as u16
