+++ wow_vanilla_messages/src/world/v1/v12/cmsg_char_delete.rs
@@ -19,7 +19,7 @@ pub struct CMSG_CHAR_DELETE {
 }
 
 impl ClientMessageWrite for CMSG_CHAR_DELETE {
-    const OPCODE: u32 = 0x38;
+    const OPCODE: u16 = 0x38;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
