+++ wow_vanilla_messages/src/world/v1/v12/cmsg_char_rename.rs
@@ -19,7 +19,7 @@ pub struct CMSG_CHAR_RENAME {
 }
 
 impl ClientMessageWrite for CMSG_CHAR_RENAME {
-    const OPCODE: u32 = 0x2c7;
+    const OPCODE: u16 = 0x2c7;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         self.size() as u16
