+++ wow_vanilla_messages/src/world/v1/v12/cmsg_destroyitem.rs
@@ -23,7 +23,7 @@ pub struct CMSG_DESTROYITEM {
 }
 
 impl ClientMessageWrite for CMSG_DESTROYITEM {
-    const OPCODE: u32 = 0x111;
+    const OPCODE: u16 = 0x111;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
